+++ phpunit/tests/fonts/font-library/wpRestFontCollectionsController.php
@@ -19,6 +19,12 @@ class Tests_REST_WpRestFontCollectionsController extends WP_Test_REST_Controller
 
 
 	public static function wpSetUpBeforeClass( WP_UnitTest_Factory $factory ) {
+		// Clear the font collections.
+		$collections = WP_Font_Library::get_instance()->get_font_collections();
+		foreach ( $collections as $slug => $collection ) {
+			WP_Font_Library::get_instance()->unregister_font_collection( $slug );
+		}
+
 		self::$admin_id  = $factory->user->create(
 			array(
 				'role' => 'administrator',
@@ -46,8 +52,6 @@ class Tests_REST_WpRestFontCollectionsController extends WP_Test_REST_Controller
 	 */
 	public function test_register_routes() {
 		$routes = rest_get_server()->get_routes();
-		$this->assertCount( 1, $routes['/wp/v2/font-collections'], 'Rest server has not the collections path initialized.' );
-		$this->assertCount( 1, $routes['/wp/v2/font-collections/(?P<slug>[\/\w-]+)'], 'Rest server has not the collection path initialized.' );
 
 		$this->assertArrayHasKey( 'GET', $routes['/wp/v2/font-collections'][0]['methods'], 'Rest server has not the GET method for collections initialized.' );
 		$this->assertArrayHasKey( 'GET', $routes['/wp/v2/font-collections/(?P<slug>[\/\w-]+)'][0]['methods'], 'Rest server has not the GET method for collection initialized.' );
@@ -115,7 +119,7 @@ class Tests_REST_WpRestFontCollectionsController extends WP_Test_REST_Controller
 		wp_set_current_user( self::$admin_id );
 		$request  = new WP_REST_Request( 'GET', '/wp/v2/font-collections/non-existing-collection' );
 		$response = rest_get_server()->dispatch( $request );
-		$this->assertErrorResponse( 'font_collection_not_found', $response, 404 );
+		$this->assertErrorResponse( 'rest_font_collection_not_found', $response, 404 );
 	}
 
 	/**